Add SkillController GET handlers

+++ management-app-service/src/main/java/com/teamcore/manageapp/service/services/skill/exceptions/SkillNotFoundException.java
@@ -2,14 +2,14 @@ package com.teamcore.manageapp.service.services.skill.exceptions;
 
 public class SkillNotFoundException extends SkillServiceException {
 
-    private int skillId;
+    private long skillId;
 
-    public SkillNotFoundException(int skillId) {
+    public SkillNotFoundException(long skillId) {
         this.skillId = skillId;
         this.message = String.format("Skill [%d] not found", skillId);
     }
 
-    public int getSkillId() {
+    public long getSkillId() {
         return skillId;
     }
 }
